requirements.txt
asgiref==3.5.1
Django==3.2.11
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
django-environ==0.8.1
psycopg2-binary==2.9.3
requests==2.27.1
